+++ src/utils/index.ts
@@ -1,0 +1,5 @@
+export { compareItems } from "./compareItems";
+export { getTreeItem } from "./getTreeItem";
+export { isAncestor } from "./isAncestor";
+export { isDroppable } from "./isDroppable";
+export { mutateTree } from "./mutateTree";